+++ src/pyscipopt/scip.pyx
@@ -533,8 +533,6 @@ cdef class Solution:
     cdef create(SCIP* scip, SCIP_SOL* scip_sol):
         if scip == NULL:
             raise Warning("cannot create Solution with SCIP* == NULL")
-        if scip_sol == NULL:
-            return None
         sol = Solution()
         sol.sol = scip_sol
         sol.scip = scip
@@ -543,11 +541,13 @@ cdef class Solution:
     def __getitem__(self, Expr expr):
         # fast track for Variable
         if isinstance(expr, Variable):
+            self._checkStage("SCIPgetSolVal")
             var = <Variable> expr
             return SCIPgetSolVal(self.scip, self.sol, var.scip_var)
         return sum(self._evaluate(term)*coeff for term, coeff in expr.terms.items() if coeff != 0)
     
     def _evaluate(self, term):
+        self._checkStage("SCIPgetSolVal")
         result = 1
         for var in term.vartuple:
             result *= SCIPgetSolVal(self.scip, self.sol, (<Variable> var).scip_var)
@@ -560,6 +560,7 @@ cdef class Solution:
         cdef SCIP_VAR* scip_var
 
         vals = {}
+        self._checkStage("SCIPgetSolVal")
         for i in range(SCIPgetNVars(self.scip)):
             scip_var = SCIPgetVars(self.scip)[i]
 
@@ -569,6 +570,12 @@ cdef class Solution:
 
             vals[name] = SCIPgetSolVal(self.scip, self.sol, scip_var)
         return str(vals)
+    
+    def _checkStage(self, method):
+        if method in ["SCIPgetSolVal", "getSolObjVal"]:
+            if self.sol == NULL and not SCIPgetStage(self.scip) == SCIP_STAGE_SOLVING:
+                raise Warning(f"{method} cannot only be called in stage SOLVING without a valid solution (current stage: {SCIPgetStage(self.scip)})")
+
 
 cdef class BoundChange:
     """Bound change."""
@@ -4151,6 +4158,7 @@ cdef class Model:
         """
         if sol == None:
             sol = Solution.create(self._scip, NULL)
+        sol._checkStage("getSolObjVal")
         if original:
             objval = SCIPgetSolOrigObj(self._scip, sol.sol)
         else: